requirements.txt
python-logstash==0.4.6
python-dotenv==0.10.3
python-dateutil==2.7.5
pika==1.1.0
boto3==1.10.7
opencv-python==4.2.0.34
Pillow==7.1.2
pdf2image==1.14.0
pytesseract==0.3.6
spacy==2.3.5
pandas==1.0.3
Unidecode==1.1.1
pymongo==3.10.1
Unidecode==1.1.1